\name{mdp_span}
\alias{mdp_span}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Evaluates the span of a vector
}
\description{
Computes the span of a vector.
}
\usage{
mdp_span(W)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{W}{
a vector.
}
}
\details{
mdp_span computes the span of the W vector: max W(s) - min W(s)
}
\value{
the value of the span of the W vector.
}
